Separate configuration from utilies
